Function pointers are not compile-time constants when using the wrong width, first stdint.h Fixes #2

+++ src/SoV_CommandLineArguments.h
@@ -19,16 +19,16 @@
 #define CLO_TYPE_PORT		3
 #define CLO_TYPE_MAC		4
 
-#define CLO_CALL(method)			(u_int)((callback_empty*)(method))
-#define CLO_SET_FIELD(type,field)	((u_int)(&(((type*)0)->field)))
+#define CLO_CALL(method)			(uintptr_t)((callback_empty*)(method))
+#define CLO_SET_FIELD(type,field)	((uintptr_t)(&(((type*)0)->field)))
 #define CLO_EOL						EOL "\t\t\t"
 
 typedef struct	commandline_option_s {
-	char		*flag;
-	char		*description;
-	u_int		type;
-	u_int		offset;
-}				commandline_option_t;
+	char			*flag;
+	char			*description;
+	u_int			type;
+	uintptr_t	offset;
+}						commandline_option_t;
 
 char **read_options(void *options, int ac, char **av, commandline_option_t *arguments);
 void print_options(commandline_option_t *arguments);
